Upgrade to VS 2019

+++ SNAPLib/Genome.cpp
@@ -242,7 +242,7 @@ Genome::loadFromFile(const char *fileName, unsigned chromosomePadding, GenomeLoc
 
     int contigNameBufferSize = 0;
     char *contigNameBuffer = NULL;
-    unsigned n;
+    size_t n;
     size_t contigSize;
     char *curName;
     for (unsigned i = 0; i < nContigs; i++) {
